\name{RImageJ-package}
\Rdversion{1.1}
\alias{RImageJ-package}
\docType{package}
\title{
R bindings for ImageJ
}
\description{
bindings between R and the java based image processing 
and analysis platform ImageJ
}
\details{
\tabular{ll}{
Package: \tab RImageJ\cr
Type: \tab Package\cr
Version: \tab 0.2-143\cr
Date: \tab 2010-02-18\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois for the R package
Wayne Rasband for ImageJ

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\references{
ImageJ: \url{http://rsbweb.nih.gov/ij/index.html}
}
\keyword{ package }
