\name{junk-package}
\Rdversion{1.1}
\alias{RImageJ-package}
\docType{package}
\title{
R bindings for ImageJ
}
\description{
bindings between R and the java based image processing 
and analysis platform ImageJ
}
\details{
\tabular{ll}{
Package: \tab junk\cr
Type: \tab Package\cr
Version: \tab 0.0-141\cr
Date: \tab 2009-06-17\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois

Maintainer: Romain Francois <francoisromain@free.fr>
}
\references{
ImageJ: http://rsbweb.nih.gov/ij/index.html
}
\keyword{ package }
